+++ internal/geogrep/main.go
@@ -20,6 +20,9 @@ func Execute(args []string) int {
 	if cfg.Command == "convert" {
 		return runConvert(cfg)
 	}
+	if cfg.Command == "list" {
+		return runList(cfg)
+	}
 	if cfg.Command == "web" {
 		return runWeb(cfg)
 	}
@@ -61,5 +64,5 @@ func Execute(args []string) int {
 
 func printUsage(parseErr error) {
 	fmt.Fprintf(os.Stderr, "error: %v\n", parseErr)
-	fmt.Fprintln(os.Stderr, "usage:\n  geogrep find [--json RESULT_PATH] [-v|--verbose[=N]] [-db|--database DB_DIR|DB_FILE] [-4 IPv4/CIDR] [-6 IPv6/CIDR6] [-d DOMAIN] [-k KEYWORD] IPv4/CIDR/IPv6/CIDR6/DOMAIN/KEYWORD\n  geogrep convert -i INPUT -o OUTPUT [--to FORMAT]\n  geogrep web [-db|--database DB_DIR|DB_FILE] [-l|--listen IP:PORT] [--webui PATH] [--api-only] [-v|--verbose[=N]]\n  geogrep version")
+	fmt.Fprintln(os.Stderr, "usage:\n  geogrep find [--json RESULT_PATH] [-v|--verbose[=N]] [-db|--database DB_DIR|DB_FILE] [-4 IPv4/CIDR] [-6 IPv6/CIDR6] [-d DOMAIN] [-k KEYWORD] IPv4/CIDR/IPv6/CIDR6/DOMAIN/KEYWORD\n  geogrep list [--json RESULT_PATH] [-db|--database DB_DIR|DB_FILE] RULESET_NAME [...]\n  geogrep convert -i INPUT -o OUTPUT [--to FORMAT]\n  geogrep web [-db|--database DB_DIR|DB_FILE] [-l|--listen IP:PORT] [--webui PATH] [--api-only] [-v|--verbose[=N]]\n  geogrep version")
 }